requirements.txt
streamlit
scikit-learn
pandas
rdkit-pypi